23.Testing with the "in-memory" Transport
From within the test, we can now fetch that transport and ask it how many messages were sent to it!

Fetching the Transport Service

Behind the scenes, every transport is actually a service in the container. Find your open terminal and run:

php bin/console debug:container async
There they are: messenger.transport.async and messenger.transport.async_priority_high. Copy the second service id.

We want to verify that the AddPonkaToImage message is sent to the transport, and we know that it's being routed to async_priority_high.

Back in the test, this is super cool: we can fetch the exact transport object that was just used from within the test by saying: $transport = self::$container->get() and then pasting the service id: messenger.transport.async_priority_high

This self:: property holds the container that was actually used during the test request and is designed so that we can fetch anything we want out of it.

Let's see what this looks like: dd($transport)

Now jump back over to your terminal and run:

php bin/phpunit

Nice! This dumps the InMemoryTransport object and... the sent property indeed holds our one message object! All we need to do now is add an assertion for this.