23.Testing with the "in-memory" Transport
Back in the test, I'm going to help out my editor by adding some inline docs to advertise that this is an InMemoryTransport. Below add $this->assertCount() to assert that we expect one message to be returned when we say $transport->... let's see... the method that you can call on a transport to get the sent, or "queued" messages is get().

Let's try it! Run:

php bin/phpunit

Got it! We're now guaranteeing that the message was sent but we've kept our tests faster and more directed by not trying to handle them synchronously. If we were using something like RabbitMQ, we also don't need to have that running whenever we execute our tests.

Next, let's talk deployment! How do we run our workers on production... and make sure they stay running?